smokeview source: add casts to eliminate warning messages
Former-commit-id: 2a0f737c1a8ead516e8de96f587cd004ebed49d6 [formerly 5b7e74f875ca13d09db791afe81a55405ba9215a]
Former-commit-id: 349c5e0ef313875a9386b0032d79927862c36306
Former-commit-id: d26a7f8ce7e118d7212a05ce49428b0e49d4a060

+++ SMV_5/source/shared/histogram.c
@@ -51,7 +51,7 @@ void init_histogram(histogramdata *histgram){
     histgram->buckets[i]=0;
   }
   histgram->ntotal=0;
-  histgram->valmin=pow(10.0,20.0);
+  histgram->valmin=(float)pow(10.0,20.0);
   histgram->valmax=-histgram->valmin;
 }
 
@@ -66,7 +66,7 @@ void copy_data2histogram(float *vals, int nvals, histogramdata *histgram){
     histgram->buckets[i]=0;
   }
   if(nvals==0){
-    valmin=pow(10.0,20.0);
+    valmin=(float)pow(10.0,20.0);
     valmax=-valmin;
   }
   else{
@@ -137,7 +137,7 @@ void merge_histogram(histogramdata *histgram1, histogramdata *histgram2){
       int ival;
 
       if(bucket1copy[i]!=0){
-        val = histgram1->valmin + (i+0.5)*dbucket1;
+        val = (float)(histgram1->valmin + (float)(i+0.5)*dbucket1);
         valmin_new=HMIN(valmin_new,val);
         valmax_new=HMAX(valmax_new,val);
         ival = (val-valmin_new)/dbucket_new;
@@ -146,7 +146,7 @@ void merge_histogram(histogramdata *histgram1, histogramdata *histgram2){
         histgram1->buckets[ival]+=bucket1copy[i];
       }
       if(histgram2->buckets[i]!=0){
-        val = histgram2->valmin + (i+0.5)*dbucket2;
+        val = (float)(histgram2->valmin + (i+0.5)*dbucket2);
         valmin_new=HMIN(valmin_new,val);
         valmax_new=HMAX(valmax_new,val);
         ival = (val-valmin_new)/dbucket_new;
